requirements.txt
python-dotenv
llama-index
llama-index-llms-groq
llama-index-llms-huggingface
llama-index-llms-ollama

llama-index-embeddings-huggingface==0.1.0
llama-index-vector-stores-chroma
vllm==0.4.0
fastapi